Split remaining endpoints
